requirements.txt
# Core Framework
streamlit>=1.35.0
streamlit-option-menu>=0.3.12
streamlit-extras>=0.4.2

# AI/ML Agents
crewai>=1.0.0
langchain>=0.2.1
langchain-google-genai>=1.0.5
langchain-community>=0.2.1
google-generativeai>=0.5.4

# Vector Database
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Topic Modeling (optional - comment out if install fails)
# bertopic>=0.16.0
# umap-learn>=0.5.6
# hdbscan>=0.8.33
scikit-learn>=1.3.0

# Data & API
arxiv>=2.1.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
networkx>=3.1
pyvis>=0.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# PDF Support
pypdf>=4.0.0
google-genai>=0.8.0
